auto generate api docs
