+++ inlang/packages/paraglide/paraglide-js/docs-api/README.md
@@ -1,0 +1,4 @@
+## Modules
+
+- [compiler-options](compiler-options.md)
+- [runtime/type](runtime/type.md)
