+++ inlang/packages/paraglide/paraglide-js/docs-api/runtime/type.md
@@ -1,0 +1,93 @@
+## Modules
+
+- [\<internal\>](-internal-.md)
+
+## Runtime
+
+> **Runtime**: `object`
+
+Defined in: [runtime/type.ts:4](https://github.com/opral/monorepo/blob/bf86cc52a6bbec968003fe3503bd3c3392eb3b48/inlang/packages/paraglide/paraglide-js/src/compiler/runtime/type.ts#L4)
+
+The Paraglide runtime API.
+
+### Type declaration
+
+#### assertIsLocale
+
+> **assertIsLocale**: [`assertIsLocale`](-internal-.md#assertislocale)
+
+#### baseLocale
+
+> **baseLocale**: [`baseLocale`](-internal-.md#baselocale)
+
+#### cookieName
+
+> **cookieName**: [`cookieName`](-internal-.md#cookiename)
+
+#### deLocalizeHref
+
+> **deLocalizeHref**: [`deLocalizeHref`](-internal-.md#delocalizehref)
+
+#### deLocalizeUrl
+
+> **deLocalizeUrl**: [`deLocalizeUrl`](-internal-.md#delocalizeurl)
+
+#### extractLocaleFromCookie
+
+> **extractLocaleFromCookie**: [`extractLocaleFromCookie`](-internal-.md#extractlocalefromcookie)
+
+#### extractLocaleFromRequest
+
+> **extractLocaleFromRequest**: [`extractLocaleFromRequest`](-internal-.md#extractlocalefromrequest)
+
+#### extractLocaleFromUrl
+
+> **extractLocaleFromUrl**: [`extractLocaleFromUrl`](-internal-.md#extractlocalefromurl)
+
+#### getLocale
+
+> **getLocale**: [`getLocale`](-internal-.md#getlocale)
+
+#### getUrlOrigin
+
+> **getUrlOrigin**: [`getUrlOrigin`](-internal-.md#geturlorigin)
+
+#### isLocale
+
+> **isLocale**: [`isLocale`](-internal-.md#islocale)
+
+#### locales
+
+> **locales**: [`locales`](-internal-.md#locales)
+
+#### localizeHref
+
+> **localizeHref**: [`localizeHref`](-internal-.md#localizehref)
+
+#### localizeUrl
+
+> **localizeUrl**: [`localizeUrl`](-internal-.md#localizeurl)
+
+#### overwriteGetLocale
+
+> **overwriteGetLocale**: [`overwriteGetLocale`](-internal-.md#overwritegetlocale)
+
+#### overwriteGetUrlOrigin
+
+> **overwriteGetUrlOrigin**: [`overwriteGetUrlOrigin`](-internal-.md#overwritegeturlorigin)
+
+#### overwriteSetLocale
+
+> **overwriteSetLocale**: [`overwriteSetLocale`](-internal-.md#overwritesetlocale)
+
+#### setLocale
+
+> **setLocale**: [`setLocale`](-internal-.md#setlocale)
+
+#### strategy
+
+> **strategy**: [`strategy`](-internal-.md#strategy)
+
+#### urlPatterns
+
+> **urlPatterns**: [`urlPatterns`](-internal-.md#urlpatterns)
