+++ inlang/packages/paraglide/paraglide-js/marketplace-manifest.json
@@ -34,7 +34,8 @@
 		},
 		"Reference": {
 			"/architecture": "./docs/architecture.md",
-			"/compiler-options": "./docs/compiler-options.md",
+			"/compiler-options": "./docs-api/compiler-options.md",
+			"/runtime": "./docs-api/runtime/-internal-.md",
 			"/errors": "./docs/errors.md"
 		},
 		"Frameworks": {
