+++ inlang/packages/paraglide/paraglide-js/package.json
@@ -29,7 +29,8 @@
 	"scripts": {
 		"dev": "tsc --watch",
 		"bench": "vitest bench --run",
-		"build": "npm run env-variables && tsc --build",
+		"build": "npm run env-variables && tsc --build && npm run generate-api-docs",
+		"generate-api-docs": "typedoc",
 		"test": "npm run env-variables && tsc --noEmit && vitest run --coverage ./src/**/*",
 		"test:watch": "npm run env-variables && vitest --watch ./src/**/*",
 		"env-variables": "node ./src/services/env-variables/create-index-file.js",
@@ -58,6 +59,9 @@
 		"memfs": "4.17.0",
 		"prettier": "^3.4.2",
 		"rolldown": "1.0.0-beta.1",
+		"typedoc": "^0.27.7",
+		"typedoc-plugin-markdown": "^4.4.2",
+		"typedoc-plugin-missing-exports": "^3.1.0",
 		"typescript": "^5.7.3",
 		"typescript-eslint": "^8.20.0",
 		"vitest": "2.1.8"
