+++ inlang/packages/paraglide/paraglide-js/src/compiler/compile-project.ts
@@ -3,7 +3,10 @@ import { selectBundleNested, type InlangProject } from "@inlang/sdk";
 import { lookup } from "../services/lookup.js";
 import { generateLocaleModules } from "./output-structure/locale-modules.js";
 import { generateMessageModules } from "./output-structure/message-modules.js";
-import { defaultCompilerOptions, type CompilerOptions } from "./compile.js";
+import {
+	defaultCompilerOptions,
+	type CompilerOptions,
+} from "./compiler-options.js";
 
 /**
  * Takes an inlang project and compiles it into a set of files.
