+++ inlang/packages/paraglide/paraglide-js/src/compiler/compile.ts
@@ -7,99 +7,10 @@ import {
 	getLocalAccount,
 	saveLocalAccount,
 } from "../services/account/index.js";
-import type { Runtime } from "./runtime/type.js";
-
-export const defaultCompilerOptions = {
-	outputStructure: "message-modules",
-	emitGitIgnore: true,
-	includeEslintDisableComment: true,
-	emitPrettierIgnore: true,
-	strategy: ["cookie", "globalVariable", "baseLocale"],
-	cookieName: "PARAGLIDE_LOCALE",
-} as const satisfies Partial<CompilerOptions>;
-
-export type CompilerOptions = {
-	/**
-	 * The path to the project to compile.
-	 *
-	 * @example
-	 *   './project.inlang'
-	 */
-	project: string;
-	/**
-	 * The path to the directory to write the output to.
-	 *
-	 * @example
-	 *   './src/paraglide'
-	 */
-	outdir: string;
-	/**
-	 * The strategy to use for getting the locale.
-	 *
-	 * The order of the strategy defines the precendence of matches.
-	 *
-	 * For example, in `['pathname', 'cookie', 'baseLocale']`, the locale will be
-	 * first tried to be detected in the pathname, then in a cookie, and finally
-	 * fallback to the base locale.
-	 *
-	 * The default ensures that the browser takes a cookie approach,
-	 * server-side takes the variable (because cookie is unvailable),
-	 * whereas both fallback to the base locale if not available.
-	 *
-	 * @default ["cookie", "variable", "baseLocale"
-	 */
-	strategy?: Runtime["strategy"];
-	/**
-	 * The name of the cookie to use for the cookie strategy.
-	 *
-	 * @default 'PARAGLIDE_LOCALE'
-	 */
-	cookieName?: string;
-	/**
-	 * Additional files that should be emmited in the outdir.
-	 *
-	 * @example
-	 *   additionalFiles: {
-	 *     "custom-file.js": "console.log('Hello, world!')"
-	 *   }
-	 */
-	additionalFiles?: Record<string, string>;
-	/**
-	 * Whether to emit a .prettierignore file.
-	 *
-	 * @default true
-	 */
-	emitPrettierIgnore?: boolean;
-
-	/**
-	 * TODO documentation
-	 */
-	urlPatterns?: Runtime["urlPatterns"];
-	/**
-	 * Whether to include an eslint-disable comment at the top of each .js file.
-	 *
-	 * @default true
-	 */
-	includeEslintDisableComment?: boolean;
-	/**
-	 * Whether to emit a .gitignore file.
-	 *
-	 * @default true
-	 */
-	emitGitIgnore?: boolean;
-	/**
-	 * The file-structure of the compiled output.
-	 *
-	 * @default "message-modules"
-	 */
-	outputStructure?: "locale-modules" | "message-modules";
-	/**
-	 * The file system to use. Defaults to `await import('node:fs')`.
-	 *
-	 * Useful for testing the paraglide compiler by mocking the fs.
-	 */
-	fs?: typeof import("node:fs");
-};
+import {
+	defaultCompilerOptions,
+	type CompilerOptions,
+} from "./compiler-options.js";
 
 // This is a workaround to prevent multiple compilations from running at the same time.
 // https://github.com/opral/inlang-paraglide-js/issues/320#issuecomment-2596951222
