+++ inlang/packages/paraglide/paraglide-js/src/compiler/index.ts
@@ -1,4 +1,7 @@
-export { defaultCompilerOptions, type CompilerOptions } from "./compile.js";
+export {
+	defaultCompilerOptions,
+	type CompilerOptions,
+} from "./compiler-options.js";
 export type { MessageBundleFunction, MessageFunction } from "./types.js";
 export type { Runtime } from "./runtime/type.js";
 export { compile } from "./compile.js";
