+++ inlang/packages/paraglide/paraglide-js/src/compiler/output-structure/message-modules.ts
@@ -4,7 +4,7 @@ import { createRuntimeFile } from "../runtime/create-runtime.js";
 import { createRegistry } from "../registry.js";
 import { jsIdentifier } from "../../services/codegen/identifier.js";
 import { escapeForSingleQuoteString } from "../../services/codegen/escape.js";
-import type { CompilerOptions } from "../compile.js";
+import type { CompilerOptions } from "../compiler-options.js";
 
 export function generateMessageModules(
 	compiledBundles: CompiledBundleWithMessages[],
