+++ inlang/packages/paraglide/paraglide-js/tsconfig.json
@@ -1,10 +1,6 @@
 {
 	"extends": "@opral/tsconfig/default",
-	"include": [
-		"src/**/*.js",
-		"src/**/*.ts",
-		"src/compiler/runtime/ambient.d.ts"
-	],
+	"include": ["src/**/*.js", "src/**/*.ts", "src/compiler/runtime/ambient.d.ts"],
 	"compilerOptions": {
 		"outDir": "./dist",
 		"rootDir": "./src",
