+++ inlang/packages/paraglide/paraglide-js/typedoc.json
@@ -1,0 +1,16 @@
+{
+	"entryPoints": [
+		"./src/compiler/compiler-options.ts",
+		"./src/compiler/runtime/type.ts"
+	],
+	"out": "./docs-api",
+	"plugin": ["typedoc-plugin-missing-exports", "typedoc-plugin-markdown"],
+	"tsconfig": "tsconfig.json",
+	"outputFileStrategy": "modules",
+	"readme": "none",
+	"excludeExternals": true,
+	"hidePageHeader": true,
+	"hideBreadcrumbs": true,
+	"hideGroupHeadings": true,
+	"hidePageTitle": true
+}
